Add Rebar in Slab Section
- Rebar material
- Rebar in section
- Flip slabs to have positive shell orientation up
- Rebar orientation

+++ script.py
@@ -39,28 +39,28 @@ if len(mdb.models.keys()) > 0:							#Deletes all other models
 
 #================ Close and delete stuff ==================#
 # This is in order to avoid corrupted files because when running in Parallels
-
-#Close and delete odb files
-import os
-import glob
-fls = glob.glob('*.odb')
-for i in fls:
-	if len(session.odbs.keys())>0:
-		session.odbs[i].close()
-	os.remove(i)
-
-
-#Delete old input
-inpt = glob.glob('*.inp')
-for i in inpt:
-	os.remove(i)
+if 0:
+	#Close and delete odb files
+	import os
+	import glob
+	fls = glob.glob('*.odb')
+	for i in fls:
+		if len(session.odbs.keys())>0:
+			session.odbs[i].close()
+		os.remove(i)
 
 
-#Delete old jobs
-jbs = mdb.jobs.keys()
-if len(jbs)> 0:
-	for i in jbs:
-		del mdb.jobs[i]
+	#Delete old input
+	inpt = glob.glob('*.inp')
+	for i in inpt:
+		os.remove(i)
+
+
+	#Delete old jobs
+	jbs = mdb.jobs.keys()
+	if len(jbs)> 0:
+		for i in jbs:
+			del mdb.jobs[i]
 
 
 
@@ -74,8 +74,8 @@ if len(jbs)> 0:
 
 runJob = 0		     	#If 1: run job
 saveModel = 0			#If 1: Save model
-cpus = 2				#Number of CPU's
-jobName = 'basicFrame'
+cpus = 8				#Number of CPU's
+jobName = 'job1'
 
 
 #================ Materials ==================#
@@ -95,6 +95,14 @@ mat2_E = 35000.0		#E-module
 mat2_v = 0.3			#Poisson
 mat2_yield = 355.0			#Yield stress
 
+# Material 3
+mat3 = "Rebar Steel"		#Material name
+mat3_Description = 'This is the description'
+mat3_dens = 8.0e-09		#Density
+mat3_E = 210000.0		#E-module
+mat3_v = 0.3			#Poisson
+mat3_yield = 355.0		#Yield stress
+
 
 #================ Parts ==================#
 #Column
@@ -105,16 +113,23 @@ imp = 0				#Initial imperfection ("triangle" shape)
 
 #Beam
 part2 = "Beam"
-sect2 = "HUP2"
+sect2 = "HEB"
 beam_len = 8000.0
 
 #Slab
 part3 = "Slab"
 sect3 = "Slab"
-deck_t = 100.0	#Thickness of slabs
+deck_t = 200.0	#Thickness of slabs
+
+#Rebars in Slab
+rebarDim = 20.0			#mm^2 diameter
+rebarArea = 3.1415*(rebarDim/2.0)**2		#mm^2
+rebarSpacing = 120.0		#mm
+rebarPosition = -80.0		#mm distance from center of section
 
 
 #================ Assembly ==================#
+#4x4  x10(5)
 x = 4			#Nr of columns in x direction
 z = 4			#Nr of columns in z direction
 y = 10			#nr of stories
@@ -147,11 +162,12 @@ static = 0					# 1 if static
 riks =   1					# 1 if Riks static
 nlg = ON					# Nonlinear geometry (ON/OFF)
 inInc = 1e-5				# Initial increment
+minInc = 1e-9
 
 #================ Loads ==================#
-LL_kN_m = -2.0	    #kN/m^2
+LL_kN_m = -2.0e1	    #kN/m^2
 
-LL=LL_kN_m * 1e-3   #N/mm^2
+LL=LL_kN_m * 1.0e-3   #N/mm^2
 
 
 
@@ -177,6 +193,16 @@ M.materials[mat2].Density(table=((mat2_dens, ), ))
 M.materials[mat2].Elastic(table=((mat2_E, mat2_v), ))
 M.materials[mat2].Plastic(table=((mat2_yield, 0.0), ))
 
+
+#================ Rebar Steel ==================#
+M.Material(description=mat3_Description, name=mat3)
+M.materials[mat3].Density(table=((mat3_dens, ), ))
+M.materials[mat3].Elastic(table=((mat3_E, mat3_v), ))
+M.materials[mat3].Plastic(table=((mat3_yield, 0.0), ))
+
+#Hardning (random linear interpolatin)
+M.materials[mat3].plastic.setValues(table=((355.0, 
+    0.0), (2000.0, 20.0)))
 
 
 
@@ -278,7 +304,13 @@ M.HomogeneousShellSection(idealization=NO_IDEALIZATION,
     poissonDefinition=DEFAULT, preIntegrate=OFF, temperature=GRADIENT, 
     thickness=deck_t, thicknessField='', thicknessModulus=None, thicknessType=
     UNIFORM, useDensity=OFF)
-
+	
+#Add rebars to section
+M.sections[sect3].RebarLayers(layerTable=(
+    LayerProperties(barArea=rebarArea, orientationAngle=0.0, barSpacing=rebarSpacing, 
+    layerPosition=rebarPosition, layerName='Layer 1', material=mat3), ), 
+    rebarSpacing=CONSTANT)	
+	
 #Create part
 M.ConstrainedSketch(name='__profile__', sheetSize= 10000.0)
 M.sketches['__profile__'].rectangle(point1=(0.0, 0.0), point2=(beam_len, beam_len))
@@ -293,6 +325,15 @@ M.parts[part3].SectionAssignment(offset=0.0,
     0.0, 0.0), ), )), sectionName='Slab', 
     thicknessAssignment=FROM_SECTION)
 
+	
+#Assign Rebar Orientation
+M.parts[part3].assignRebarOrientation(
+    additionalRotationType=ROTATION_NONE, axis=AXIS_1, fieldName='', localCsys=
+    None, orientationType=GLOBAL, region=Region(
+    faces=M.parts[part3].faces.findAt(((0.1, 
+    0.1, 0.0), (0.0, 0.0, 1.0)), )))
+
+	
 #Create surface
 #Gets name Slab_A1-1.Surf
 M.parts[part3].Surface(name='Surf', side2Faces=
@@ -344,19 +385,20 @@ for a in range(len(alph)-1):
 				vector=(x_d*a , col1_height*(e+1), z_d*n))
 
 #Beams in z (numb) direction
-for a in range(len(alph)-0):
+#a=0
+for a in [0,x-1]:
 	for n in range(len(numb)-1):
 		for e in range(len(etg)):
 			inst = part2+"_"+ alph[a]+numb[n] + "-" + alph[a]+numb[n+1] + "-"+etg[e]
 			beamList.append(inst)
-			#import and name instance
+			# import and name instance
 			M.rootAssembly.Instance(dependent=dep,name=inst, part=M.parts[part2])
-			#Rotate instance
+			# Rotate instance
 			M.rootAssembly.rotate(angle=-90.0, axisDirection=(
 				0.0,1.0, 0.0), axisPoint=(0.0, 0.0, 0.0), instanceList=(inst, ))
-			#Translate instance in x,y and z
+			# Translate instance in x,y and z
 			M.rootAssembly.translate(instanceList=(inst, ),
-				vector=(x_d*a , col1_height*(e+1), z_d*n))
+				vector=(x_d*a , col1_height*(e+1), z_d*n))	
 
 
 #================ Slabs ==================#
@@ -366,11 +408,10 @@ for a in range(len(alph)-1):
 			inst = part3+"_"+ alph[a]+numb[n] + "-"+etg[e]
 			slabList.append(inst)
 			M.rootAssembly.Instance(dependent=dep,name=inst, part=M.parts[part3])
-			M.rootAssembly.rotate(angle=90.0, axisDirection=(
+			M.rootAssembly.rotate(angle=-90.0, axisDirection=(
 							1.0,0.0, 0.0), axisPoint=(0.0, 0.0, 0.0), instanceList=(inst, ))
 			M.rootAssembly.translate(instanceList=(inst, ),
-							vector=(x_d*a,col1_height*(e+1),z_d*n))
-
+							vector=(x_d*a,col1_height*(e+1),z_d*(n+1)))
 
 
 
@@ -429,8 +470,8 @@ if static:
 	M.StaticStep(description='description', 
 		initialInc=inInc, name=stepName, nlgeom=nlg, previous=oldStep)
 elif riks:
-	M.StaticRiksStep(description='description', 
-		initialArcInc=inInc, name=stepName, nlgeom=nlg, previous=oldStep, maxLPF=1.0)
+	M.StaticRiksStep(description='description', initialArcInc=inInc,
+		name=stepName, nlgeom=nlg, previous=oldStep, maxLPF=1.0, minArcInc=minInc)
 
 
 
@@ -480,7 +521,7 @@ for a in range(len(alph)-1, 0,-1):
 				((a*x_d, (e+1)*col1_height, n*z_d), ), )), userMode=DOF_MODE_MPC, userType=0)
 
 #Column to beam in z(num) direction
-for a in range(len(alph)):
+for a in [0,x-1]:
 	for n in range(len(numb)-1):
 		for e in range(len(etg)):
 			col = part1+"_"+ alph[a]+numb[n] + "-" +etg[e]
@@ -496,7 +537,7 @@ for a in range(len(alph)):
 				((a*x_d, (e+1)*col1_height, n*z_d), ), )), userMode=DOF_MODE_MPC, userType=0)
 
 #Column to beam in negative z(num) direction
-for a in range(len(alph)):
+for a in [0,x-1]:
 	for n in range(len(numb)-1,0,-1):
 		for e in range(len(etg)):
 			col = part1+"_"+ alph[a]+numb[n] + "-" +etg[e]
@@ -543,7 +584,7 @@ for a in range(len(alph)-1):
 				((x_d*a+1 , col1_height*(e+1), z_d*n), ), ), name=inst+'_surf')
 
 #Create beam surfaces in z (numb) direction
-for a in range(len(alph)-0):
+for a in [0, x-1]:
 	for n in range(len(numb)-1):
 		for e in range(len(etg)):
 			inst = part2+"_"+ alph[a]+numb[n] + "-" + alph[a]+numb[n+1] + "-"+etg[e]
@@ -560,9 +601,35 @@ for a in range(len(alph)-1):
 				M.rootAssembly.instances[inst].edges.findAt(
 				((x_d*a+1, col1_height*(e+1), z_d*n), ),
 				((x_d*a+1, col1_height*(e+1), z_d*n+x_d), ),
-				((x_d*a, col1_height*(e+1), z_d*n+1), ),
-				((x_d*a+x_d, col1_height*(e+1), z_d*n+1), ), ))
+				#((x_d*a, col1_height*(e+1), z_d*n+1), ),
+				#((x_d*a+x_d, col1_height*(e+1), z_d*n+1), ), 
+				))
 
+# a=0
+# for n in range(len(numb)-1):
+	# for e in range(len(etg)):
+		# inst = part3+"_"+ alph[a]+numb[n] + "-"+etg[e]
+		# M.rootAssembly.Surface(name=inst+'_edges', side1Edges=
+			# M.rootAssembly.instances[inst].edges.findAt(
+			# ((x_d*a+1, col1_height*(e+1), z_d*n), ),
+			# ((x_d*a+1, col1_height*(e+1), z_d*n+x_d), ),
+			# ((x_d*a, col1_height*(e+1), z_d*n+1), ),
+			# #((x_d*a+x_d, col1_height*(e+1), z_d*n+1), ),
+			# ))
+
+# a=x-2
+# for n in range(len(numb)-1):
+	# for e in range(len(etg)):
+		# inst = part3+"_"+ alph[a]+numb[n] + "-"+etg[e]
+		# M.rootAssembly.Surface(name=inst+'_edges', side1Edges=
+			# M.rootAssembly.instances[inst].edges.findAt(
+			# ((x_d*a+1, col1_height*(e+1), z_d*n), ),
+			# ((x_d*a+1, col1_height*(e+1), z_d*n+x_d), ),
+			# #((x_d*a, col1_height*(e+1), z_d*n+1), ),
+			# ((x_d*a+x_d, col1_height*(e+1), z_d*n+1), ),
+			# ))
+				
+				
 #Join beam surfaces to match slabs
 for a in range(len(alph)-1):
 	for n in range(len(numb)-1):
@@ -570,14 +637,42 @@ for a in range(len(alph)-1):
 			inst = part3+"_"+ alph[a]+numb[n] + "-"+etg[e]
 			beam1 = part2+"_"+ alph[a]+numb[n] + "-" + alph[a+1]+numb[n] + "-"+etg[e]
 			beam2 = part2+"_"+ alph[a]+numb[n+1] + "-" + alph[a+1]+numb[n+1] + "-"+etg[e]
-			beam3 = part2+"_"+ alph[a]+numb[n] + "-" + alph[a]+numb[n+1] + "-"+etg[e]
-			beam4 = part2+"_"+ alph[a+1]+numb[n] + "-" + alph[a+1]+numb[n+1] + "-"+etg[e]
+			#beam3 = part2+"_"+ alph[a]+numb[n] + "-" + alph[a]+numb[n+1] + "-"+etg[e]
+			#beam4 = part2+"_"+ alph[a+1]+numb[n] + "-" + alph[a+1]+numb[n+1] + "-"+etg[e]
 			M.rootAssembly.SurfaceByBoolean(name=inst+'_beamEdges', 
 				surfaces=(
 				M.rootAssembly.surfaces[beam1+'_surf'], 
 				M.rootAssembly.surfaces[beam2+'_surf'],
-				M.rootAssembly.surfaces[beam3+'_surf'],
-				M.rootAssembly.surfaces[beam4+'_surf']))
+				#M.rootAssembly.surfaces[beam3+'_surf'],
+				#M.rootAssembly.surfaces[beam4+'_surf']
+				))
+
+# a=0
+# for n in range(len(numb)-1):
+	# for e in range(len(etg)):
+		# inst = part3+"_"+ alph[a]+numb[n] + "-"+etg[e]
+		# beam1 = part2+"_"+ alph[a]+numb[n] + "-" + alph[a+1]+numb[n] + "-"+etg[e]
+		# beam2 = part2+"_"+ alph[a]+numb[n+1] + "-" + alph[a+1]+numb[n+1] + "-"+etg[e]
+		# beam3 = part2+"_"+ alph[a]+numb[n] + "-" + alph[a]+numb[n+1] + "-"+etg[e]
+		# M.rootAssembly.SurfaceByBoolean(name=inst+'_beamEdges', 
+			# surfaces=(
+			# M.rootAssembly.surfaces[beam1+'_surf'], 
+			# M.rootAssembly.surfaces[beam2+'_surf'],
+			# M.rootAssembly.surfaces[beam3+'_surf'],))
+
+# a=x-2
+# for n in range(len(numb)-1):
+	# for e in range(len(etg)):
+		# inst = part3+"_"+ alph[a]+numb[n] + "-"+etg[e]
+		# beam1 = part2+"_"+ alph[a]+numb[n] + "-" + alph[a+1]+numb[n] + "-"+etg[e]
+		# beam2 = part2+"_"+ alph[a]+numb[n+1] + "-" + alph[a+1]+numb[n+1] + "-"+etg[e]
+		# beam4 = part2+"_"+ alph[a+1]+numb[n] + "-" + alph[a+1]+numb[n+1] + "-"+etg[e]
+		# M.rootAssembly.SurfaceByBoolean(name=inst+'_beamEdges', 
+			# surfaces=(
+			# M.rootAssembly.surfaces[beam1+'_surf'], 
+			# M.rootAssembly.surfaces[beam2+'_surf'],
+			# M.rootAssembly.surfaces[beam4+'_surf'],))
+
 
 #Tie slabs to beams (beams as master)
 for a in range(len(alph)-1):
@@ -645,7 +740,7 @@ mdb.Job(atTime=None, contactPrint=OFF, description='', echoPrint=OFF,
     explicitPrecision=SINGLE, getMemoryFromAnalysis=True, historyPrint=OFF, 
     memory=90, memoryUnits=PERCENTAGE, model=modelName, modelPrint=OFF, 
     multiprocessingMode=DEFAULT, name=jobName, nodalOutputPrecision=SINGLE, 
-    numCpus=cpus, numDomains=2, numGPUs=0, queue=None, resultsFormat=ODB, scratch=
+    numCpus=cpus, numDomains=cpus, numGPUs=0, queue=None, resultsFormat=ODB, scratch=
     '', type=ANALYSIS, userSubroutine='', waitHours=0, waitMinutes=0)
 
 if runJob == 1:
